# /results page is slow, this PR perf profile helps diagnose that and makes it faster