those code produces the following output:

```

pulled alumni in 0.0 seconds (total: 0.0 sec)
pulled count_* in 0.01 seconds (total: 0.01 sec)
pulled orgs in 0.0 seconds (total: 0.01 sec)
pulled Stats1 in 0.15 seconds (total: 0.17 sec)
pulled Stats2 in 0.17 seconds (total: 0.34 sec)
pulled bounty_history in 0.08 seconds (total: 0.42 sec)
pulled completion_rate in 0.01 seconds (total: 0.43 sec)
pulled universe_total_usd in 0.01 seconds (total: 0.44 sec)
pulled bounty_average_turnaround in 0.61 seconds (total: 1.05 sec)
pulled bounty_median_pickup_time in 0.59 seconds (total: 1.64 sec)
pulled final in 0.0 seconds (total: 1.64 sec)


```

and introduces caching of that information