@owocki 2018-06-16
@mbeacom ideally id like tot take any with more than 0.1 response and wrap them in a redis cache.  whats the easiest way to do this with our current tooling?  i could also just create a DBCache table if we don't want to intro the redis dependancy