@mbeacom 2018-06-18
@owocki I think the Redis cache would be the way to go.