@owocki 2018-06-18
@mbeacom i dont want to introduce more complexity into our app just for this one page... is redis already on the roadmap? what is your recommended django cache management class?