@owocki 2018-06-29
@mbeacom 

ee090ecb37ecd79ea2a0a59467d274fa77907fa2

takes the page load time down from 4 seconds to 500 ms:

<img width="485" alt="screen shot 2018-06-29 at 3 44 03 pm" src="https://user-images.githubusercontent.com/513929/42116145-4742b390-7bb3-11e8-983e-a016012a8385.png">

<img width="490" alt="screen shot 2018-06-29 at 3 43 49 pm" src="https://user-images.githubusercontent.com/513929/42116147-477161ae-7bb3-11e8-9ab2-43f19c58ff30.png">

its using 

```
CACHES = {'default': env.cache()}
```

mind if we 🚢 this now and then move it over to redis at some later date in the future?